refactor: 악곡 랭킹 테이블 컴포넌트 구조 분리
